Trying to fix type problem for gpu implementation

+++ Python/nn.py
@@ -23,7 +23,7 @@ display_step = 100
 x_all = np.loadtxt('../data/train.csv',
                delimiter = ',',
                skiprows = 1,
-               dtype = np.float32,
+               dtype = np.int64,
                usecols = range(0, n_features + 1))
 
 # NEED TO GENERATE RANDOM INDEXES INSTEAD
